backend/requirements.txt
# asgiref==3.8.1
# Django==5.0.4
# django-filter==24.2
# djangorestframework==3.15.1
# Markdown==3.6
# mysqlclient==2.2.4
# PyJWT==2.8.0
# sqlparse==0.5.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
channels==4.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
Markdown==3.6
multidict==6.0.5
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
websockets==12.0
yarl==1.9.4
